Cleanup code in preparation to re-enable webpack

+++ src/hn-api.ts
@@ -2,41 +2,35 @@ import {APIGatewayProxyResult} from "aws-lambda";
 import firebase from "firebase/app";
 import "firebase/database";
 import DataSnapshot = firebase.database.DataSnapshot;
+import Reference = firebase.database.Reference;
+import App = firebase.app.App;
 
 const API_URL = "https://hacker-news.firebaseio.com";
 const API_VERSION = "/v0";
 
 export async function hn(): Promise<APIGatewayProxyResult> {
-    console.log("starting");
-    const app = firebase.initializeApp({databaseURL: API_URL});
-    const db = app.database().ref(API_VERSION);
-    // console.dir(db);
-    const path = "item/192327";
-    // const promise = new Promise(
-    //     (resolve: (value: any) => void, reject: (reason: any) => void): void => {
-    //         db.child(path).once(
-    //             "value",
-    //             (snapshot: DataSnapshot) => {
-    //                 console.log(snapshot);
-    //                 resolve(snapshot.val());
-    //             },
-    //             reject
-    //         );
-    //     }
-    // );
-    const item = db.child(path);
-    // console.dir(item);
-    console.log("fetching");
-    const snapshot = await item.once("value");
-    // console.dir(snapshot.exists());
-    const data = snapshot.val();
-    console.dir(data);
-    await app.delete();
+    let app: App;
+    let data: any;
+
+    try {
+        // TODO - Look for an existing Firebase application before creating a new one.
+        app = firebase.initializeApp({databaseURL: API_URL});
+        const db: Reference = app.database().ref(API_VERSION);
+        const path = "item/192327";
+        const item = db.child(path);
+        const snapshot: DataSnapshot = await item.once("value");
+        data = snapshot.val();
+    } catch (error) {
+        console.error(error);
+    } finally {
+        if (app) {
+            // TODO - Identify a more appropriate way to manage Firebase resources/applications.
+            await app.delete();
+        }
+    }
+
     return {
         statusCode: 200,
-        body: JSON.stringify({
-            here: "in here",
-            data: data,
-        }),
+        body: JSON.stringify(data),
     };
 }
